refactor(app): move reaction/species picker subsystem into cs/picker/
Moves the self-contained reaction/species picker components
(cs-set-filter, swr-reaction-picker(-shell), swr-state-list,
swr-state-select, state-select, latex-select) out of the flat shared/
folder into a new cs/picker/ folder, since they are only used together
and by cs/swr-filter-component.tsx and the author set edit page's
reaction-builder.tsx. Updates all import paths accordingly. No
behavioral changes.
